pyproject.toml
[project]
name = "comicrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "accelerate>=1.10.1",
    "attrdict>=2.0.1",
    "chroma>=0.2.0",
    "chromadb>=1.3.5",
    "einops>=0.8.1",
    "langchain>=0.3.27",
    "langchain-experimental>=0.3.4",
    "numpy>=2.0.2",
    "ollama>=0.6.1",
    "open-clip-torch>=3.2.0",
    "openai>=2.8.1",
    "pillow>=11.3.0",
    "sentence-transformers>=5.1.2",
    "torch>=2.2.2",
    "transformers>=4.57.3",
]

[tool.uv.extra-build-dependencies]
autoawq = ["torch"]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
# Define the custom index URL
[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"


